refactor: migrate item collection to TypeScript
